# [feat] Refactor front-End error messaging

+++ frontend/src/app/services/hotels.service.ts
@@ -1,7 +1,8 @@
-import { inject, Injectable, signal, WritableSignal } from "@angular/core";
+import { effect, inject, Injectable, signal, WritableSignal } from "@angular/core";
 import { MatDialog } from '@angular/material/dialog';
 import { HttpClient } from '@angular/common/http';
 import { Observable, of, tap } from 'rxjs';
+import { AuthService } from './auth.service';
 
 export type Hotel = {
   id?: string;
@@ -18,12 +19,24 @@ export type Hotel = {
 export class HotelsService {
   protected readonly dialog = inject(MatDialog);
   private readonly http = inject(HttpClient);
+  private readonly authService = inject(AuthService);
   private readonly url = `http://localhost:8082/api/v1/hotel`;
 
   public readonly allHotels: WritableSignal<Hotel[]> = signal<Hotel[]>([]);
+  public readonly user = this.authService.user;
+
+  constructor() {
+    effect(() => {
+      if (!this.user()) {
+        this.allHotels.set([]);
+      }
+    });
+  }
 
   // === GET ALL HOTELS === //
   public getAll(): Observable<Array<Hotel>> {
+    if (!this.user()) return of();
+
     if (this.allHotels().length) {
       return of(this.allHotels());
     }
@@ -42,6 +55,8 @@ export class HotelsService {
 
   // === CREATE === //
   public add(request: Hotel): Observable<Hotel> {
+    if (!this.user()) return of();
+
     const { name, title, city, address, distance } = request;
     return this.http.post<Hotel>(this.url, {
       name,
@@ -58,6 +73,8 @@ export class HotelsService {
 
   // === UPDATE === //
   public update(request: Hotel): Observable<Hotel> {
+    if (!this.user()) return of();
+
     const { id, name, title, city, address, distance } = request;
     return this.http.put<Hotel>(`${this.url}/${id}`, {
       name: name.trim(),
@@ -80,6 +97,8 @@ export class HotelsService {
 
   // === DELETE === //
   public delete(request: Hotel): Observable<void> {
+    if (!this.user()) return of();
+
     const { id } = request;
     return this.http.delete<void>(`${this.url}/${id}`).pipe(
       tap(() => this.allHotels.update(hotels => hotels.filter(hotel => hotel.id !== id)))
